Added/Fixed font weights 400-700 & their italics
Both of our fonts can now use any combination of
font weight 400-700 with or without italics. This is important for when
we import Rich Text from our CMS, for example, in ServiceDetail.jsx for
each package text loaded inside <PortableText> component. Now italics
will be italics and bolds will be bold

- Also previously we weren't even loading the right 1 variaton of
fonts because the url was mistakenly not pointing to any file skull
emoji